Move generate_compiler_repro_string to after_aot

[ghstack-poisoned]
